Move auth utils inline
